requirements.txt
Django~=2.2.5
django-multiselectfield~=0.1.8
psycopg2
selenium
pytesseract
Pillow
xlsxwriter
celery
django-celery-beat
django-celery-results
pylibmc
sqlalchemy
